Refactor to move userenv and admin config to Git::Hooks.
The configuration variables userenv and admin, which were available in the
check-acls and check-jira plugins were promoted to the "githooks" configuration
section. Along with them, the functions grok_userenv, match_user, and im_admin
were also migrated to the Git::Hooks module so that they can be used by any
access control plugin.

The old configuration variables in the plugins are still supported but are
now documented as deprecated.