Remove all state variables
In several places we used state variables, trying to be efficient.
But this makes the code non-reentrant, which is a no-no now that
we turned the plugins into modules. So, I implemented a new method
Git::More::cache to allow plugins to keep their cached data inside
the Git::More object. Using this I was able to remove all uses of
state variables.