CheckJira: change meaning of CheckJira.by-assignee option
This is a non-compatible change.

Previously, the by-assignee option required the name of an environment
variable. But this is duplicating poorly the functionality of the
githooks.userenv global option. So, I changed its meaning to now
require a boolean (integer). If true it now checks if the
authenticated_user (as infered by the githooks.userenv option) is
the current issue's assignee.