Make JIRA::Client be recommended instead of required
JIRA::Client is required to use the CheckJira plugin, but it's not
required to Git::Hooks per se and for other plugins. So, it's now
only recommended. Interestingly, this change does not affect the
plugin's test, because it already used a fake JIRA::Client class.

Issue: https://rt.cpan.org/Ticket/Display.html?id=82167