Substitute Git for App::gh::Git
The Git.pm module is distributed with git and is not at CPAN yet. A
while back I decided to require App::gh::Git, which is a copy of Git
maintained by the GitHub people. However, I needed to require App::gh
instead of App::gh::Git directly, due to some versioning problems in
the latter, which isn't good, because App::gh carries some heavy
dependencies of its own.

In the end, I decided to get back to use Git directly. I can't depend
on it directly, but this way our dependencies are fewer and I have at
least the peace of mind of being using The Real Thing (c).