Try to find Git if it's not on @INC
Since I can't depend directly on Git I made Makefile.PL try to require
it to see if it can find it. If it can't, we don't install Git::Hooks
and produce a verbose message telling the user why and what she can do
about it.

Moreover, we allow the user to specify a directory for Git.pm using
the GITPERLLIB environment variable. This is inspired by the way the
git Perl scripts do it, but it's not quite the same. The git Perl
scripts prepend the directories in GITPERLLIB to @INC, making them
used to every use/require statement in the program, which can lead to
problems. In our case, we make a temporary change to @INC just in
order to require Git.pm.

If Makefile.PL can't require Git.pm directly, it tries to find it in
some usual places and inform the user before aborting the
installation.

Some of this was inspired by Randall Schwartz's post at:
http://git.661346.n2.nabble.com/better-way-to-find-Git-pm-officially-td7416362.html