Execute tests in a temporary directory
The test harness was being executed without changing directories. This
was risky because any bug in a test could result (as it in fact
resulted) in messing with the author's own Git::Hooks repository.

Now we do a chdir at the start of every test to a temporary
directory.

This change triggered an error in the t/02-gerrit-change-id.t test as
it tried to find the gerrit-commig-msg.sh script in a directory
relative to the current directory, which isn't know. So, I moved the
script to t/02-gerrit-change-id.t's own __DATA__ section and the
script is copied to a file in the temporary directory at the start of
every test, so that its path is relative to a know path.