Prepare t/test-functions.pl for rebase testing
The pre-rebase hook is invoked without the GIT_DIR environment
variable and with cwd pointing to the repo's wc_path. We have to
detect this in the hook script in order to define GIT_CONFIG
accordingly.

Define a remote called 'clone' in the 'repo' repository to make it
easier to invoke git push in the tests.

Define a new sub test_ok_match to be used by the post-commit hook.