Collect errors from all hooks before dying
Git::Hooks::run_hooks now invokes every internal and external hook
inside an eval block in order to detect if they fail. After invoking
all of them it dies if at least one failed.

Moreover, hooks now shouldn't die directly when they detect a
problem. Instead, they should invoke Git::More::error to produce the
error message and return a boolean value indicating if they have
succeeded. This way they don't need to stop at the first error, but
can detect many more and generate more information to the user.

In order to support old hooks, run_hooks checks their return
values. If they return a boolean, they're treated as modern hooks. If
they return undef (and $@ is empty, so that it wasn't an exception)
they are considered to have succeeded.

Currently, when a commit is rejected due to a failed check in a
plugin, the user gets an error message. If it is an inconsequential
problem the user might decide to bypass the commit checks by passing
the B<-n> option to disable the hooks altogether. Or he may disable
the specific plugin by defining an environment variable. Either way,
the commit might have other problems that got undetected simply
because the hook aborted after detecting the first one. When the user
disables all or some checks he may well be letting pass other errors.

Yet another argument for this new mechanism is that it is necessary if
we want to implement Gerrit's asynchronous hooks.