refactor: convert a global variable into a cached value
CheckJira kept a JIRA::Client object in a global variable. This change
gets rid of that variable and uses a slot in the plugin's Git::More
cache instead. It also uses the new Git::More::clean_cache method to
delete the JIRA::Client object at the end of the hook, forcing a
disconnection from JIRA when it's no longer needed.