Search for a useable Git.pm during build
Makefile.PL now tries to find a useable Git module if it can't find it
in @INC, by going to some usual places. If it can find it it creates a
file t/GITPERLLIB which is used by t/test-functions.pl during the
tests to set the GITPERLLIB environment variable.

This way we can hopefully test and install the distribution, even
though the user might have to set its own GITPERLIB or PERL5LIB
variables later to use it.