Git::More: fix config grokking on Windows
The Git module isn't as well supported on Windows as it is on Unix. In
particular, the Git::command* functions don't cope well with changes
in the global $/ after they're called and before the command output is
read. Here I localize $/ before invoking the Git routine.

I also substitute Git::command for Git::command_output_pipe in
get_config to make the code simpler.