Support Windows by invoking hooks indirectly
It's difficult to use a Perl script directly as a hook on Windows
because of the way that system (mis-)handles the shee-bang. Also,
since Windows don't support symlinks it's not possible to configure
hooks the way we do on Unix.

So, here I configure each hook as a sh script which invokes the Perl
script indirectly.

I do this in the test framework but I don't document it, as we're not
officially supporting Windows yet.