Fix commit grokking on pushing a new branch
The Git::More::get_commits method was invoking git-rev-list with wrong
arguments when it got a null SHA-1 as $old_commit. This is a know
problem discussed, eg, at http://stackoverflow.com/questions/4523668.

Now, in this situation it knows how to prepare a suitable argument
list for git-rev-list.

This bug was reported by Marcelo Bissaro.

Signed-off-by: Gustavo L. de M. Chaves <gnustavo@cpan.org>