Fix commit grokking on deleting a branch
The Git::More::get_commits method was invoking git-rev-list with wrong
arguments when it got a null SHA-1 as $new_commit, which happens when
one pushes to delete a remote branch.

Now, in this situation it returns an empty list, meaning that no new
commit has been pushed.

This bug was reported by Marcelo Bissaro.

Signed-off-by: Gustavo L. de M. Chaves <gnustavo@cpan.org>