Make tests independent of git's global init.templatedir option
After I started using a personal repository directory template for
myself, by configuring init.templatedir, the tests began to fail. This
is because my default template already installs some hooks which
conflicts with the ones that the testing framework tries to install.

So, now, during the tests, I create a standard template directory and
use it in all git-init and git-clone commands.

Signed-off-by: Gustavo L. de M. Chaves <gnustavo@cpan.org>