Make Git::More::read_commit_msg_file stripspace directly
The method Git::More::read_commit_msg_file used to invoke the external
command 'git stripspace -s' to cleanup the commit message. However, it
used Git::command_bidi_pipe to do it but it seems that this method
isn't working on FreeBSD. So, I reimplemented the cleanup funcionality
in Perl.

Signed-off-by: Gustavo L. de M. Chaves <gnustavo@cpan.org>