CheckJira: make githooks.checkjira.project strict
When this option was used, JIRAs for non-specified projects were simply
ignored, which isn't the best semantics. The user could be led to
believe that she was citing correct JIRAs when in reality they could not
even exist.

There are two other possible semantics for this option: it could either
reject any JIRA of a non-specified project or accept them if they
respected all other CheckJira checks.

It's not obvions which one is better but in the end I decided for the
more strict semantics, viz. to reject JIRAs of non-specified
projects. One argument for this is that JIRAs should respect ALL checks,
and the githooks.checkjira.project is such a check and should be
respected too. Another argument is that if the user wants to cite JIRAs
of other projects she is able to do so by citing them outside of the
place defined by the githooks.checkjira.matchlog option, since those
would not even be detected.

Note that this is a backwards-incompatible change. But it's really minor
and I don't think it'll have an impact. And if it has, it will be
for the better, I think.