Capture external hooks output
Since we didn't care about external hooks's output, their STDOUT and
STDERR were simply going to the default stream. This is OK for standard
Git hooks but not for Gerrit hooks, because the output ended up only on
Gerrit's error log in the server.

Now we redirect STDOUT and STDERR to a temporary file before invoking
external hooks and capture anything left there later so that we can send
it to the normal output and also insert it in any Gerrit comment we
generate along with error messages.