Fix merge commit checks
The CheckLog and CheckJira plugins were invoking Git::More::get_commits
to grok commit information in the commit-msg, patchset-created, and
draft-published hooks. However, as it is currently implemented
get_commits can't grok a single commit when it's a merge commit, because
the command "git rev-list ^NEW^ NEW" isn't enough to exclude all parents
of NEW.

In this I implemented another method Git::More::get_commit to grok a
single commit and changed CheckLog and CheckJira to use it.