Substitute filter_files_in_* for get_diff_files
The method Git::Hooks::get_diff_files returned a hash mapping filenames
to change statuses. It was being used so far only in the CheckStructure
plugin. There were two problems with it.

First, it was unecessarily complicated because it returned a hash and
only the filenames were being used. So, it would be simpler to return a
list of file names.

Second, it wasn't general enough. As it were, it could only grok the
list of files changed between two commits, which we use in the update
and the pre-receive hook. However, the pre-commit hook needs to grok
changes in the index, and Gerrit's patchset-created hook needs to grok
changes in a single commit.

So, in this commit I remove the get_diff_files method and substituted
for it three new ones:

- filter_files_in_index: returns the files changed in the index with
  respect to HEAD.

- filter_files_in_range: returns the files changed between two commits.

- filter_files_in_commit: returns the files changed in a single commit
  with respect to its parents.

All three return lists of file names.