CheckJira: interpolate $+ into regexp fixversions too
The githooks.checkjira.fixversion option special feature allowing one to
use capture groups in the BRANCH argument and interpolate the matching
text into the FIXVERSION argument was implemented only for non-regexp
FIXVERSIONs. The restriction was intended to avoid dangerously
interpolating regexp meta-characters into another regexp.

But it was an oversight, since we can simply escape-quote the matched
text before interpolating it into a regexp FIXVERSION. Which is exactly
what this commit implements.