CheckFile: replace temp filename with real name in command output
The output from the external commands may refer to the filename which
they were called to check. Since the actual file is a temporary its name
will confound the user. So, we replace any mention of the temporary
filename in the output with the name of the file in the repository.