Fix cleverness in error detection
Mark VanderBrink brought the problem with the following construct to
my attention:

  <loop> {
    <expression>
        or $errors++
            and next;
    <loop body>
  }

Since $errors is 0 to begin with, the first time <expression>
evaluates to false "$errors++" will also evaluate to false, which will
shortcut and won't execute "next" as it was supposed to.

The solution is to replace "$errors++" by "++$errors" making it true
in the first evaluation.

I admit that this construct is a little bit "too clever" to be
warranted. However, I'll stick to it for the time being.