Replace every postfix++ by ++prefix
The postfix ++ operator is more "magic" than the prefix ++ operator in
that it evaluates to the value of the variable "before" the expression
instead of after it.

Most of the time they're exchangeable because they're used just for
the side effect and not for the result. However, since the prefix one
is simpler to understand I'm replacing every instance of the postfix
operator by the prefix one.