Use dzil's MetaProvides::Package
CPANTS [1] complained that "this distribution does not have a list of
provided modules defined in META.yml" and suggested the use of
Dist:Zilla's MetaProvides plugin.

The plugin complained that lib/Git/Hooks/TUTORIAL.pod didn't have a
name nor an abstract [2] which was solved by specifying them at the
start of the file.

Ref: [1] http://cpants.cpanauthors.org/dist/Git-Hooks
Ref: [2] http://blogs.perl.org/users/polettix/2011/11/distzilla-podweaver-and-bin.html