Insert README.pod in INSTALL.SKIP
Karen Etheridge explained to me [0] that I should not install
README.pod because it's installed in the wrong place. So, I'm
excluding it in the INSTALL.SKIP file.

I took the chance to also move the TODO.pod exclusion from the
MANIFEST.SKIP to the INSTALL.SKIP file so that the files are part of
the distribution but aren't installed. This way both can be seen on
MetaCPAN.

Ref: [0] https://github.com/gnustavo/Git-Hooks/issues/10