Make CheckJira require JIRA::REST instead of use it
Since we're commited to simply suggest instead of require the
JIRA::REST module it stands to reason that no Git::Hooks module can
'use' it directly. Instead we 'require' it inside an eval block so
that we can provide a helpful error message to the user.

I took the chance to change the error message in CheckLog regarding a
missing Text::SpellChecker module so that they have the same wording.