Turn Gerrit::REST into a mere recommendation
Gerrit::REST isn't required unless the user is using one of Gerrit's
hooks. So, it doesn't make a lot of sense to make it a requirement of
Git::Hooks, just a recommendation.

Moreover, Gerrit::REST depends on REST::Client which is difficult to
install in some environments.

Issue: https://github.com/gnustavo/Git-Hooks/issues/11