Ignore changes on submodule paths
The Git::More filter_files_in_* methods return lists of files changed
in a commit or a range of commits. When submodules are used the names
of their paths were being returned in the list, which caused problems
when they were being fed to the git-cat-file command. So, now we pass
the option --ignore-submodules to the git-diff-* commands so that the
module paths aren't returned anymore.