Make t/02-check-commit.t tests idempotent
The check_cannot_push routine was leaving the current repository with
a new commit which could not have been pushed to the clone. That's why
after every call to it we needed to invoke "git reset --hard"
explicitly. Now the routine invokes it itself, making it easier to use
it.