Replace githooks.checkcommit.check-ref by ...check-code
This is an incompatible change! I'm replacing the
githooks.checkcommit.check-ref option by another one called
githooks.checkcommit.check-code, which performs the checks for every
commit in an update or pre-receive hook, not for every affected
reference. Now it applies also to the local pre-commit hook.

The githooks.checkcommit.check-ref option was ill-conceived. The
CheckCommit plugin is meant to check "commits", not "references". As
such, all checks should apply to the local commit being constructed in
a pre-commit hook and to all commits being pushed in an update or
pre-receive hook.

The callback routines associated with the option now get three
arguments ($git, $commit, $ref) instead of two ($git, $ref).