Fix Git::More::get_commits for new branches
When a new branch is created and is pushed before receiving any
commits the get_commits routine was miscalculating its commit
range. The problem was that it was excluding from the 'git
for-each-ref ref/heads/' command any commits equal to the new branch's
$new_commit. But for new branches with no new commits $new_commit is
surely reachable by other branches from which the new one was created.

I don't know why I was excluding $new_commit from the list using
grep. Perhaps I thought that would improve performance a little...

Anyway, the fix could be simply the removal of the grep operation on
the list of commits. But I've been shown that there is an easier and
faster way to calculate the @exclude list in [0]. The fix made the
code a lot cleaner too.

Ref: [0] https://github.com/gnustavo/Git-Hooks/issues/24