Fix Git::More::get_commits for all branches
For previously existing branches we were obtaining the list of new
commits in get_commits using "git rev-list $new_commit
^$old_commit". However, if there is any merge commit along the path
from $old_commit to $new_commit that command also returned the commits
along the other side of the merge, which we don't want to get.

So, we always have to get the commits with "git rev-list --not
--all". If $old_commit is not the NULL commit we also exclude it, even
though I guess it would be already excluded by '--not --all'.

I've been refered to this problem in [0].

Issue: [0] https://github.com/gnustavo/Git-Hooks/issues/26