Generalizes voting mechanism for Gerrit
The existing configuration options to specify which votes are cast to
the Gerrit's change (githooks.gerrit.review-label,
githooks.gerrit.vote-nok, and githooks.gerrit.vote-ok) allow for
voting in a single label. Sometimes it's useful to be able to vote in
multiple labels. So, this commit introduces two new configuration
options and deprecates the existing ones.

The new options are called githooks.gerrit.votes-to-approve and
githooks.gerrit.votes-to-reject. Each one may specify multiple votes
in different labels.