Fix Git::More::get_commits in post-* hooks
In a pre-receive or pre-update hook the affected refs point to the OLD
commit. In this case, 'git-rev-parse --not --all' returns the negated
OLD commit as we expect to exclude all commits leading to it.

However, in a post-receive or post-update hook, the affected refs
already point to the NEW commit. So, we have to remove it form the
list returned by 'git-rev-parse --not --all' and insert a negated OLD
commit instead. Otherwise, the get_commits routine always returns an
empty list of commits in the post-* hooks.

Issue: https://github.com/gnustavo/Git-Hooks/issues/27