CheckJira: Change githooks.checkjira.comment specification
Now one must set this option to 'all' to make the comment not
visibility restricted. Letting it empty was being considered as not
being set.

This should be a non backwards compatible change but for the fact that
comment insertion wasn't really working up to now. So, I'm considering
it a fix.