Fix regression on Git::More::get_commits
v1.13.1 introduced a regression that was previously fixed on
v1.12.2. More specifically, the bug was introduced on b1e9105d.

v1.12.2 fixed the get_commits method when it was invoked as a
pre-receive/update hook for a new branch with no commits in it.

v1.13.1 fixed get_commits when it was invoked as a
post-receive/post-update hook, but in doing so it broke the fix
implemented in v1.12.2.

To be able to fix both issues get_commits must be aware of which hook
it's being called in. In order to do so I'm introducing a new method
Git::More::hookname which returns the name of the current hook.

Issue: https://github.com/gnustavo/Git-Hooks/issues/31