Use Git::Repository as the underlying git wrapper module
This is a big and incompatible change, which will ultimately lead to a
major version increment.

The "official" Git module is distributed along with git itself but it
hasn't changed much in the last years. On the other hand, there are
several good Git-wrapper modules at CPAN. I evaluated them a while ago
and convinced myself that Git::Repository is one of the best, being
well maintained and offering an interesting plugin architecture that I
can leverage to reorganize the Git::Hooks distribution.

So, this is the first of a series of commits in which I'll make
Git::Repository the underlying git wrapper module for Git::Hooks.

In this commit the major changes are these:

* Transforms Git::More into Git::Repository::Plugin::GitHooks.

* Transforms Git::More::Message into Git::Message.

* Move most of the utility functions defined in Git::Hooks into
  Git::Repository::Plugin::GitHooks.

Ref0: https://metacpan.org/pod/Git