CheckStructure: remove plugin
The CheckStructure plugin's features can be implemented by a
combination of the CheckFile and CheckReference
plugins. CheckStructure's configuration syntax require one to specify
the repository structure by a usually large Perl data strucuture,
which can be a problem for the Perl-challenged people.

Moreover, CheckStructure is the only plugin using the
Git::Repository::Plugin::GitHooks::eval_gitconfig routine, which
allows one to evaluate unconstrained Perl code from the configuration
file. Since this can be easily turned into a security hole I'm going
to get rid of that method in a following commit.