Add the G::R::P::G::prepare_hook and load_plugins methods
All the routines used to prepare specific routines which were on the
Git::Hooks module were moved to G::R::P::GitHooks and a new method was
created so that G::H::run_hooks can invoke it to prepare the hooks.

This allowed us to get rid of the G::R::P::GitHooks::hookname method
too.