Substitute Try::Tiny for Error
The Error module is no longer recommended since 2009-12-19 [0], but we
continued using it because we dependend on the Git module, which still
depended on Error. Since we no longer depend on Git I'm replacing
Error by Try::Tiny, which is a more modern exception handling module.

Ref0: https://metacpan.org/changes/distribution/Error