Fix G::R::P::GitHooks::filter_files_in_range method
This method is invoked with two commits. When it was invoked with the
first commit being the undefined commit it tried to compare the empty
tree with the second commit's tree. However, the first commit is
undefined when a new branch or tag is pushed. In this case, the method
returned every file in the repository, which, for large repositories
led to a very long processing time.

What we really want in this case, though is to grok the same list of
commits as returned by the get_commits method, and then compare the
first commit's parent tree with the last commit's tree. This way we
can really grok the list of files changed by the new commits pointed
by the new branch or tag.

However, even in this case there is a special case when the list's
first commit is a root commit, i.e., when it has no parent. In this
case we don't want to return all files in the repository because it
can be huge. Instead, we simply return the empty list.