Notify: send one email per matching branch+rule
Currently we collect all the recipients of all matching rules in order
to send a single notification. However, this notification may be complex
because it may contain changes made in different branches. And the
recipients may have a hard time figuring out why they're being notified.

So, now we send a specific email to each matching rule, to its specific
recipients. Each email contains the changes made to a single branch. So,
we may have to send several emails, but they will be shorter and easier
to understand.