Notify: send one email per rule
Lots of changes in this one.

Beginning with the configuration, we replaced the 'include' option by
the 'rule' option, which can receive a list of pathspecs to filter the
commits using the same syntax as the git-log command. This allows us to
send one email per 'rule' configuration, and the emails show only the
commits affecting the rule's pathspecs.

In order to do that I had to remove the caching performed by the
get_affected_ref_commits method, because it didn't take into account the
new arguments (specifically, $paths) it can get.

One cannot filter by reference anymore. All requests I got for this
plugin required only filters by file and not by reference. So, in order
to simplify the configuration I'm dropping support for this.

I'm no longer using git-diff to check which files changed because I'm no
longer using regexen, but pathspecs to filter. So, I'm invoking a
git-log for each affected reference, and for each rule. So, I'm able to
use the get_affected_ref_commits method to grok commits, and I can
format the notification body explicitly.

The transport configuration now allows one to use any of the
Email::Sender::Transport::* modules, even passing arguments to
them. Usually, though, the default transport is enough and it doesn't
have to be configured explicitly.

There are seven tests working.