Notify: put more information in the notification
Now the notification email contains a block of information between the
preamble and the log, telling the context of the notification in a
format like this:

  REPOSITORY: myproject
  BRANCH: refs/heads/master
  CHANGED BY: username
  FROM: 75550b66ab08536787487545904fb062c6e38a7f
  TO:   6eaa6a84fbd7e2a64e66664f3d58707618e20c72
  FILTER: lib/Git/Hooks/

The FILTER line appears only if the rule that generated the notification
contained one or more pathspecs.

Also, the notification's subject was changed a bit.