Do not check $ENV{HOME} in get_config
The get_config method used to check if the HOME environment variable was
set because for a long time I thought that Gerrit invoked its hooks with
it undefined. However, recently I discovered that in fact this was a
side-effect of the way a colleague of mine used to start our Gerrit
server.

Hence, the message get_config showed when it detected that HOME wasn't
set was misleading.

This commit removes the check from the get_config method.