CheckCommit: fake missing environment variables
When we fake a commit during the check_pre_commit routine we have to
grok the author's and committer's name and email. We try to grok them
from Git's environment variables, but in some cases not all of them are
set.

So, we have to detect when they're not set and fake them too.