Replace error/get_errors by fault/get_faults
The G::R::P::GitHooks::error method accepted three positional
parameters (prefix, message, details), the last one being optional. The
first parameter was the prefix, which is usually the package name of the
plugin invoking the method.

The new G::R::P::GitHooks::fault method replaces it to make it easier to
invoke. The new method make the prefix optional and, in its absense,
groks it by invoking the caller funtion. This way one does not have to
pass the package name explicitly in the majority of the fault() calls,
as it was necessary with the error() calls.

The new method signature is (message, info), where message is a string
and info an optional hash-ref which may contain a 'prefix' and/or some
'details'.