Move documentation from GitHub wiki to POD files
The README and the Tutorials were being kept in the GitHub wiki, but
this is suboptimal because it's difficult to keep them in sync with the
code evolution.

I got inspiration from the Minilla [0] distribution and internalized
that documentation as POD files in the our distribution.

Ref0: https://github.com/tokuhirom/Minilla