CheckFile: add new option deny-case-conflict
This new option rejects commits which add files with names that would
conflict with each other or with the names of other files already in the
repository in case-insensitive filesystems, such as the ones on Windows.

There are several examples of pre-commit hooks implementing this. I was
inspired mainly by [0].

Ref0: https://github.com/pre-commit/pre-commit-hooks