Reformat error messages
The 'fault' method now accepts more contextual information via its
hash-ref argument, such as a commit SHA-1, a reference name, and a
configuration option name. This information is now presented in a
"context" line before the actual error message and the details.

This allowed me to simplify several error messages, because they don't
need to mention that contextual information explicitly, since it will
appear already at the context line.

The new format is like this:

  [Git::Hooks::CheckFile: option sizelimit, at commit 7d76af34]

  The file 'bigfile' is too big.

  It has 1000000 bytes but the current limit is 100000 bytes.
  Please, check your configuration option.