CheckFile: fix test for case-sensitive filesystems
The test suite has two tests that should be performed only on
case-sensitive filesystems. We were skipping them only on Windows, but
Darwin also has case-insensitive filesytems.

Now we don't check by the OS name, but we actually check if the
filesystem is case-sensitive or not in order to skip or to perform the
tests.