CheckFile: Fix remote hooks
In remote hooks (pre-receive, update, etc.) the CheckFile plugin was
calculating the set of affected files using git-diff between the old and
new commit of each affected reference. However, this is not correct in
every situation. For instance, when one pushes a merge commit, the diff
will include all files changed in the source branch, even if its commits
were already in the server.

The correct way is to calculate the affected files for each new commit
pushed. And for merge commits, one should consider only files that were
conflicted, because non-conflicted files were already checked when we
processed the commit in which they were originally changed.