CheckFile: Fix 'deny-token' option during a pre-commit hook
On a pre-commit hook the deny_token routine invokes the git-diff-index
command to find invalid tokens. It wasn't passing the --cached option to
the command, which is necessary for it to diff agains the index. Without
the option the command was diffying against the working tree.

Issue: https://github.com/gnustavo/Git-Hooks/issues/48