Add Test::UnusedVars dzil plugin
It tests if there are unused variables.

It found a few of them which were fixed.

The $decoration variable, in PrepareLog.pm, wasn't really unused, but
the tests warned as if it were. I opened a issue on the Test::Vars
module [0] reporting a possible false positive. I changed the code to
avoid this problem. Fortunately, the result is shorter and simpler
than before.

Ref: https://github.com/houseabsolute/p5-Test-Vars/issues/36