Fix the RequireUseStrict/RequireUseWarnings issues
When dzil builds the distribution its PkgVersion plugin inserts a line
containing an assignment after the package line of every module. Since
this line appears before the 'use strict' and the 'use warnings' lines
Perl::Critict complains about two rules.

There is a pending issue to resolve it in Perl::Critic [0] but while
it's not integrated we have to solve it differently. This commit moves
the 'use strict' and the 'use warnings' lines of every module to the
start of the file to avoid those two rules.

However, this triggers another Perl::Critic rule, viz.,
Modules::RequireExplicitPackage. Fortunately this rule can be
configured to disregard the usage of specific modules such as strict
and warnings, which is configured in the perlcritic.rc file.

Ref: https://github.com/rjbs/Dist-Zilla/pull/168