Fix awk-emulation splits
In several places I was using "split / /" to split a string in
whitespaces. However, this actually splits it in "every" space
character, not exactly what I wanted.

I wanted the awk-emulation split behaviour documented in "perldoc -f
split", which splits the string using a sequence of one of more
whitespaces as delimiters. For that the correct incantation is:

  split ' '