PrepareLog: fix tests
The tests defined a GIT_EDITOR which had to effectively edit the commit
message in order to make it non empty. We used a perl script which
edited the file in place. However, this began to fail on FreeBSD for
some reason I don't understand.

This commit changes the git-commit commands to insert a commit message
using the -m option. So, the GIT_EDITOR doesn't need to edit the file
anymore.