GitHooks: Fix pre-receive hook when invoked by Bitbucket PRs
After upgrading Bitbucket's External Hooks plugin [0], from version
4.3-1 to 4.5-2, the pull-request "merge" button in the UI became
greyed-out and its tooltip shows this error message:

  Git::Repository::Plugin::GitHooks: get_affected_refs():
  no affected refs set at /usr/local/share/perl/5.18.2/Git/Hooks.pm line 65.

The problem is that when Bitbucket renders the PR page it invokes the
pre-receive hook without any input. The GitHooks::_get_affected_ref
routine croaked in this situation. But it really didn't need to
bother. It should simply return a reference to an empty hash, so that
all the Git::Hooks plugins would simply skip any checks.

I took the chance to make a small change in how the _prepare_input_data
routine logs information about the pre-receive input. Now it logs the
input as an array of lines as read from STDIN. Previously it logged the
data structure parsed from the lines. Also, now it logs it at level
'info' instead of 'debug', because this information is in the same level
as the arguments the hooks get from the command line.

Ref0: https://marketplace.atlassian.com/apps/1211631/external-hooks