Add a way to access githooks.userenv value
Config item githooks.userenv is used to name an environmental variable
which provides the authenticated user name. This is normally
USER when accessing repo via SSH or REMOTE_USER when using
a webserver.
When it is not possible to just name an environmental variable,
the value can be generated by a perl sentence.
In such a case, the value should be available for use in
other configurations, e.g. CheckReference acls.

Signed-off-by: Mikko Koivunalho <mikko.koivunalho@iki.fi>