Make all modules use 5.016 instead of 5.010
Git::Hooks should support the oldest Perl version shipped with the
oldest Ubuntu or CentOS distribution still supported.

CentOS 6 is reaching EOL on December. So, Git::Hooks doesn't need to
support Perl 5.010 anymore. The oldest distro to watch now is CentOS 7,
which uses Perl 5.16.3 and is going to be supported until June of 2024.

Since the 'strict' pragma is enabled by default when we 'use 5.016' we
removed all their explicit uses.