CheckFile: deal with filenames containing unusual characters
In the output of git-log, pathnames with "unusual" characters are quoted
as explained for the configuration variable core.quotePath (see
git-config(1)): "by enclosing the pathname in double-quotes and escaping
those characters with backslashes in the same way C escapes control
characters (e.g.  \t for TAB, \n for LF, \\ for backslash) or bytes with
values larger than 0x80 (e.g. octal \302\265 for "micro" in UTF-8)."