Suppress git warnings when they're expected
In some places we invoke Git::Repository::Run with the expectation that
the git command will fail. We already suppress the exception by passing
the option {fatal => [-129, -128]} to it. However, the warnings printed
to STDERR are still shown to the user, which we don't want.

In order to suppress the warnings, we add the option {quiet => 1} to
these invocations.