Add timeout support
Adds a new configuration option (githooks.timeout) which sets a time
limit to how log the hook can run. If it runs longer than the specified
time, in seconds, it aborts.

It is used by the new Git::Repository::Plugins::GitHooks::check_timeout
method, which is invoked repeatedly in all the loops which may take an
unbound amount of time.

Also, the new Git::Repository::Plugins::GitHooks::fail_on_faults method
is used to abort the hook when any faults have been registered so far.